Migrating from JDB to JPS (Hibernate)

+++ src/main/resources/scripts.sql
@@ -1,17 +1,17 @@
 
 create table friend (
-	id integer NOT NULL AUTO_INCREMENT,
+	id bigint NOT NULL AUTO_INCREMENT,
 	name varchar(100) NOT NULL,
 	age integer,
 	primary key(id)
 );
 
 create table address(
-	id integer not null auto_increment, 
+	id bigint not null auto_increment, 
 	location varchar(20), 
 	street_name varchar(100) not null, 
 	city varchar(50) not null, 
-	friend_id integer, 
+	friend_id bigint, 
 	primary key(id),
     foreign key (friend_id) references friend(id)
 );
